jenkinsctl packaging and refactoring
