+++ jenkinsctl/MANIFEST.in
@@ -1,0 +1,3 @@
+include requirements-common.txt
+include requirements-dev.txt
+recursive-include jenkinsctl/templates *
