+++ jenkinsctl/README.md
@@ -1,0 +1,8 @@
+# Jenkinsctl
+
+A CLI and python library used to create Jenkins jobs for different predefined
+flows.
+
+## Steps to build and publish
+
+Place an aws credentials file in `./aws_credentials`, and do a `docker build`.
