+++ jenkinsctl/docker/build_install.sh
@@ -1,0 +1,7 @@
+#!/bin/ash
+
+echo "local install"
+python3 setup.py install --force
+
+echo "build wheel"
+python3 setup.py bdist_wheel -d dist
