+++ jenkinsctl/docker/run_tests.sh
@@ -1,0 +1,4 @@
+#!/bin/ash
+
+pycodestyle *.py
+pylint *.py jenkinsctl
