+++ jenkinsctl/jenkinsctl/definitions.py
@@ -1,0 +1,5 @@
+import os
+
+ROOT_DIR = os.path.dirname(os.path.dirname(__file__))
+
+TEMPLATES_FOLDERS_DIR = os.path.join(ROOT_DIR, 'jenkinsctl/templates')
