+++ jenkinsctl/jenkinsctl/flows/__init__.py
@@ -1,0 +1,5 @@
+from jenkinsctl.flows.abstract_flow import AbstractFlow
+from jenkinsctl.flows.generic_app_flow import GenericAppFlow
+from jenkinsctl.flows.kubernetes_deployer_flow import KubernetesDeployerFlow
+
+__all__ = ["AbstractFlow", "GenericAppFlow", "KubernetesDeployerFlow"]
