+++ jenkinsctl/jenkinsctl/templates/generic-app-flow-trigger.xml.j2
