+++ jenkinsctl/jenkinsctl/templates/k8s-deployer-trigger.xml.j2
