+++ jenkinsctl/jenkinsctl/utils/__init__.py
@@ -1,0 +1,10 @@
+# from os.path import dirname, basename, isfile
+# import glob
+#
+# modules = glob.glob(dirname(__file__)+"/*.py")
+# __all__ = [basename(f)[:-3] for f in modules if isfile(f) and not f.endswith('__init__.py')]
+
+from jenkinsctl.utils.jenkins_custom import JenkinsCustom
+from jenkinsctl.utils.job_template import JobTemplate
+
+__all__ = ["JenkinsCustom", "JobTemplate"]
