+++ jenkinsctl/requirements-dev.txt
@@ -1,0 +1,5 @@
+-r requirements-common.txt
+
+pycodestyle==2.3.1
+pylint==1.7.4
+s3pypi==0.5.0
