+++ jenkinsctl/setup.cfg
@@ -1,0 +1,2 @@
+[pycodestyle]
+max-line-length = 100
